Address review comments and fix pre-commit errors
