+++ experimental/__init__.py
@@ -3,8 +3,6 @@
 # This source code is licensed under the BSD license found in the
 # LICENSE file in the root directory of this source tree.
 
-__version__ = "0.1.4"
-
 ################################################################################
 # Import most common subpackages
 ################################################################################
